TESS_Terminal_Configurable/requirements.txt
# TESS Terminal - Configurable Edition
# Full Feature Set

# === Core Dependencies ===
groq>=0.4.0
openai>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
rich>=13.0.0

# === Memory & Vector DB ===
chromadb>=0.4.0
sentence-transformers>=2.0.0  # For embeddings

# === Web & Browser ===
# Option 1: Playwright (Recommended - modern, reliable)
playwright>=1.40.0
# Run after install: playwright install chromium

# Option 2: Selenium (Legacy fallback)
selenium>=4.0.0
webdriver-manager>=4.0.0

requests>=2.28.0
beautifulsoup4>=4.11.0

# === System & Files ===
pillow>=10.0.0          # Screenshots
pyautogui>=0.9.0        # GUI automation (optional)
schedule>=1.2.0         # Task scheduling
watchdog>=3.0.0         # File watching (Librarian)

# === File Conversion ===
python-docx>=0.8.0      # DOCX handling
docx2pdf>=0.1.0         # DOCX to PDF (Windows/Mac)
PyPDF2>=3.0.0           # PDF handling

# === Document AI (NEW) ===
PyPDF2>=3.0.0           # PDF text extraction
pytesseract>=0.3.10     # OCR for images
# Note: Tesseract OCR engine must be installed separately
# Windows: https://github.com/UB-Mannheim/tesseract/wiki
# Mac: brew install tesseract
# Linux: sudo apt install tesseract-ocr

# === Google APIs ===
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# === Utilities ===
python-dotenv>=1.0.0
colorama>=0.4.0
